refactor: Move relocation processing to elf.rs
