+++ libwild/src/elf.rs
@@ -42,6 +42,7 @@ use crate::platform::FrameIndex;
 use crate::platform::ObjectFile;
 use crate::platform::Platform;
 use crate::platform::RawSymbolName as _;
+use crate::platform::Relaxation as _;
 use crate::platform::Relocation;
 use crate::platform::RelocationSequence;
 use crate::platform::SectionFlags as _;
@@ -88,6 +89,7 @@ use linker_utils::elf::secnames::*;
 use linker_utils::elf::shf;
 use linker_utils::elf::sht;
 use linker_utils::elf::stt;
+use linker_utils::relaxation::RelocationModifier;
 use linker_utils::utils::read_string;
 use linker_utils::utils::read_u32;
 use linker_utils::utils::read_uleb128;
@@ -108,6 +110,7 @@ use std::io::Read as _;
 use std::mem::offset_of;
 use std::num::NonZeroU32;
 use std::ops::Range;
+use std::sync::atomic;
 use std::sync::atomic::Ordering;
 use zerocopy::FromBytes;
 use zerocopy::IntoBytes;
@@ -429,7 +432,8 @@ impl platform::Platform for Elf {
     ) -> Result {
         match state.relocations(section.index)? {
             RelocationList::Rela(relocations) => {
-                state.load_section_relocations::<A, Rela>(
+                load_section_relocations::<A, Rela>(
+                    state,
                     common,
                     queue,
                     resources,
@@ -439,7 +443,8 @@ impl platform::Platform for Elf {
                 )?;
             }
             RelocationList::Crel(relocations) => {
-                state.load_section_relocations::<A, Crel>(
+                load_section_relocations::<A, Crel>(
+                    state,
                     common,
                     queue,
                     resources,
@@ -463,7 +468,8 @@ impl platform::Platform for Elf {
     ) -> Result {
         match state.relocations(section.index)? {
             RelocationList::Rela(relocations) => {
-                state.load_debug_relocations::<A, Rela>(
+                load_debug_relocations::<A, Rela>(
+                    state,
                     common,
                     queue,
                     resources,
@@ -473,7 +479,8 @@ impl platform::Platform for Elf {
                 )?;
             }
             RelocationList::Crel(relocations) => {
-                state.load_debug_relocations::<A, Crel>(
+                load_debug_relocations::<A, Crel>(
+                    state,
                     common,
                     queue,
                     resources,
@@ -1836,7 +1843,7 @@ fn process_eh_frame_relocations<'data, 'scope, A: Arch<Platform = Elf>, R: Reloc
 
                 // We currently always load all CIEs, so any relocations found in CIEs always need
                 // to be processed.
-                layout::process_relocation::<A, <R::Sequence<'data> as RelocationSequence>::Rel>(
+                process_relocation::<A, <R::Sequence<'data> as RelocationSequence>::Rel>(
                     object,
                     common,
                     rel,
@@ -1940,7 +1947,7 @@ fn process_section_exception_frames<'data, 'scope, A: Arch<Platform = Elf>, R: R
         // section.
         if let Some(eh_frame_section) = object.format_specific.eh_frame_section {
             for rel in frame_data.relocations.rel_iter() {
-                layout::process_relocation::<A, <R::Sequence<'data> as RelocationSequence>::Rel>(
+                process_relocation::<A, <R::Sequence<'data> as RelocationSequence>::Rel>(
                     object,
                     common,
                     &rel,
@@ -3892,4 +3899,250 @@ impl platform::BuiltInSectionDetails for BuiltInSectionDetails {}
 pub(crate) struct DynamicSymbolDefinitionExt {
     pub(crate) hash: u32,
     pub(crate) version: u16,
+}
+
+fn load_section_relocations<'scope, 'data, A: Arch<Platform = Elf>, R: Relocation>(
+    state: &layout::ObjectLayoutState<'data, Elf>,
+    common: &mut CommonGroupState<'data, Elf>,
+    queue: &mut layout::LocalWorkQueue,
+    resources: &'scope layout::GraphResources<'data, '_, Elf>,
+    section: layout::Section,
+    relocations: impl Iterator<Item = R>,
+    scope: &Scope<'scope>,
+) -> Result {
+    let mut modifier = RelocationModifier::Normal;
+    for rel in relocations {
+        if modifier == RelocationModifier::SkipNextRelocation {
+            modifier = RelocationModifier::Normal;
+            continue;
+        }
+        modifier = process_relocation::<A, R>(
+            state,
+            common,
+            &rel,
+            state.object.section(section.index)?,
+            resources,
+            queue,
+            false,
+            scope,
+        )
+        .with_context(|| {
+            format!(
+                "Failed to copy section {} from file {state}",
+                layout::section_debug::<Elf>(state.object, section.index)
+            )
+        })?;
+    }
+
+    Ok(())
+}
+
+fn load_debug_relocations<'scope, 'data, A: Arch<Platform = Elf>, R: Relocation>(
+    state: &layout::ObjectLayoutState<'data, Elf>,
+    common: &mut CommonGroupState<'data, Elf>,
+    queue: &mut layout::LocalWorkQueue,
+    resources: &'scope layout::GraphResources<'data, '_, Elf>,
+    section: layout::Section,
+    relocations: impl Iterator<Item = R>,
+    scope: &Scope<'scope>,
+) -> Result {
+    for rel in relocations {
+        let modifier = process_relocation::<A, R>(
+            state,
+            common,
+            &rel,
+            state.object.section(section.index)?,
+            resources,
+            queue,
+            true,
+            scope,
+        )
+        .with_context(|| {
+            format!(
+                "Failed to copy section {} from file {state}",
+                layout::section_debug::<Elf>(state.object, section.index)
+            )
+        })?;
+        ensure!(
+            modifier == RelocationModifier::Normal,
+            "All debug relocations must be processed"
+        );
+    }
+
+    Ok(())
+}
+
+#[inline(always)]
+fn process_relocation<'data, 'scope, A: Arch, R: Relocation>(
+    object: &layout::ObjectLayoutState<'data, A::Platform>,
+    common: &mut CommonGroupState<'data, A::Platform>,
+    rel: &R,
+    section: &<A::Platform as Platform>::SectionHeader,
+    resources: &'scope layout::GraphResources<'data, '_, A::Platform>,
+    queue: &mut layout::LocalWorkQueue,
+    is_debug_section: bool,
+    scope: &Scope<'scope>,
+) -> Result<RelocationModifier> {
+    let args = resources.symbol_db.args;
+    let mut next_modifier = RelocationModifier::Normal;
+    if let Some(local_sym_index) = rel.symbol() {
+        let symbol_db = resources.symbol_db;
+        let local_symbol_id = object.symbol_id_range.input_to_id(local_sym_index);
+        let symbol_id = symbol_db.definition(local_symbol_id);
+        let mut flags = resources.local_flags_for_symbol(symbol_id);
+        flags.merge(resources.local_flags_for_symbol(local_symbol_id));
+        let rel_offset = rel.offset();
+        let r_type = rel.raw_type();
+        let section_flags = <A::Platform as Platform>::section_flags(section);
+
+        let rel_info = if let Some(relaxation) = A::new_relaxation(
+            r_type,
+            object.object.raw_section_data(section)?,
+            rel_offset,
+            flags,
+            symbol_db.output_kind,
+            section_flags,
+            true,
+            None,
+        )
+        .filter(|relaxation| args.should_relax() || relaxation.is_mandatory())
+        {
+            next_modifier = relaxation.next_modifier();
+            relaxation.rel_info()
+        } else {
+            A::relocation_from_raw(r_type)?
+        };
+
+        let section_is_writable = section.is_writable();
+        let mut flags_to_add = layout::resolution_flags(rel_info.kind);
+
+        if !section_flags.is_alloc() {
+            // Non-alloc sections never get dynamic relocations, so there's nothing to do here.
+        } else if rel_info.kind.is_tls() {
+            if does_relocation_require_static_tls(rel_info.kind) {
+                resources
+                    .has_static_tls
+                    .store(true, atomic::Ordering::Relaxed);
+            }
+
+            if layout::needs_tlsld(rel_info.kind)
+                && !resources.uses_tlsld.load(atomic::Ordering::Relaxed)
+            {
+                resources.uses_tlsld.store(true, atomic::Ordering::Relaxed);
+            }
+        } else if flags_to_add.needs_direct() && flags.is_interposable() {
+            if section_is_writable {
+                common.allocate(part_id::RELA_DYN_GENERAL, elf::RELA_ENTRY_SIZE);
+            } else if flags.is_function() {
+                // Create a PLT entry for the function and refer to that instead.
+                flags_to_add.remove(ValueFlags::DIRECT);
+                flags_to_add |= ValueFlags::PLT | ValueFlags::GOT;
+            } else if !flags.is_absolute() {
+                match args.copy_relocations_enabled() {
+                    crate::args::CopyRelocations::Allowed => {
+                        flags_to_add |= ValueFlags::COPY_RELOCATION;
+                    }
+                    crate::args::CopyRelocations::Disallowed(reason) => {
+                        // We don't at present support text relocations, so if we can't apply a copy
+                        // relocation, we error instead.
+                        bail!(
+                            "Direct relocation ({}) to dynamic symbol from non-writable section, \
+                            but copy relocations are disabled because {reason}. {}",
+                            A::rel_type_to_string(r_type),
+                            resources.symbol_debug(symbol_id),
+                        );
+                    }
+                }
+            }
+        } else if flags.is_ifunc()
+            && rel_info.kind == RelocationKind::Absolute
+            && section_is_writable
+            && symbol_db.output_kind.is_relocatable()
+        {
+            common.allocate(part_id::RELA_DYN_GENERAL, elf::RELA_ENTRY_SIZE);
+        } else if symbol_db.output_kind.is_relocatable()
+            && rel_info.kind == RelocationKind::Absolute
+            && flags.is_address()
+        {
+            if section_is_writable {
+                common.allocate(part_id::RELA_DYN_RELATIVE, elf::RELA_ENTRY_SIZE);
+            } else if !is_debug_section {
+                bail!(
+                    "Cannot apply relocation {} to read-only section. \
+                    Please recompile with -fPIC or link with -no-pie",
+                    A::rel_type_to_string(r_type),
+                );
+            }
+        }
+
+        // For ifunc symbols with GOT-relative references (like R_X86_64_GOTPCRELX), we need a
+        // separate GOT entry for address equality. The main GOT entry will be used by the PLT stub
+        // with an IRELATIVE relocation, while this extra entry will contain the PLT stub address so
+        // that all references to the ifunc return the same address.
+
+        let relocation_needs_got = flags_to_add.needs_got();
+
+        if flags.is_ifunc() && !symbol_db.output_kind.is_static_executable() {
+            flags_to_add |= ValueFlags::GOT | ValueFlags::PLT;
+        }
+
+        if flags.is_ifunc() && relocation_needs_got && !symbol_db.output_kind.is_relocatable() {
+            flags_to_add |= ValueFlags::IFUNC_GOT_FOR_ADDRESS;
+        }
+
+        let atomic_flags = &resources.per_symbol_flags.get_atomic(symbol_id);
+        let previous_flags = atomic_flags.fetch_or(flags_to_add);
+
+        if !previous_flags.has_resolution() {
+            if flags.is_ifunc() && symbol_db.output_kind.is_static_executable() {
+                atomic_flags.fetch_or(ValueFlags::GOT | ValueFlags::PLT);
+            }
+
+            queue.send_symbol_request::<A>(symbol_id, resources, scope);
+            // If the canonical symbol belongs to a shared object, use the file ID of symbols that
+            // made the reference.
+            let undefined_check_file_id = symbol_db.file_id_for_symbol(if flags.is_dynamic() {
+                local_symbol_id
+            } else {
+                symbol_id
+            });
+            if layout::should_emit_undefined_error::<A::Platform>(
+                object.object.symbol(local_sym_index)?,
+                object.file_id,
+                undefined_check_file_id,
+                flags,
+                args,
+                symbol_db.output_kind,
+            ) {
+                let symbol_name = symbol_db.symbol_name_for_display(symbol_id);
+                let source_info =
+                    A::get_source_info(object.object, &object.relocations, section, rel_offset)
+                        .context("Failed to get source info")?;
+
+                if args.should_error_on_unresolved_symbols() {
+                    resources.report_error(error!(
+                        "Undefined symbol {symbol_name}, referenced by {}\n    {}",
+                        source_info, object.input,
+                    ));
+                } else {
+                    crate::error::warning(&format!(
+                        "Undefined symbol {symbol_name}, referenced by {}\n    {}",
+                        source_info, object.input,
+                    ));
+                }
+            }
+        }
+
+        if flags_to_add.needs_copy_relocation() && !previous_flags.needs_copy_relocation() {
+            queue.send_copy_relocation_request::<A>(symbol_id, resources, scope);
+        }
+    }
+    Ok(next_modifier)
+}
+
+/// Returns whether the supplied relocation type requires static TLS. If true and we're writing a
+/// shared object, then the STATIC_TLS will be set in the shared object which is a signal to the
+/// runtime loader that the shared object cannot be loaded at runtime (e.g. with dlopen).
+fn does_relocation_require_static_tls(rel_kind: RelocationKind) -> bool {
+    layout::resolution_flags(rel_kind) == ValueFlags::GOT_TLS_OFFSET
 }
